Formated code files. Implemented some functions.
