+++ code/wallet.py
@@ -1,0 +1,35 @@
+from cryptography.hazmat.backends import default_backend
+from cryptography.hazmat.primitives.asymmetric import rsa
+from cryptography.hazmat.primitives import serialization
+
+class Wallet:
+    def __init__(self):
+        self.private_key, self.public_key = self.generate_wallet()
+        self.coins = 0
+
+    def generate_wallet(self):
+        private_key = rsa.generate_private_key(
+            public_exponent=65537,
+            key_size=2048,
+            backend=default_backend()
+        )
+
+        public_key = private_key.public_key()
+
+        '''
+        Serialized keys are byte objects, we cant sign or verify with them, maybe delete them?
+                 
+        # Serialize private key
+        pem_private_key = private_key.private_bytes(
+            encoding=serialization.Encoding.PEM,
+            format=serialization.PrivateFormat.PKCS8,
+            encryption_algorithm=serialization.NoEncryption()
+        )
+
+        # Serialize public key
+        pem_public_key = public_key.public_bytes(
+            encoding=serialization.Encoding.PEM,
+            format=serialization.PublicFormat.SubjectPublicKeyInfo
+        )
+        '''
+        return private_key, public_key
